Rendered stars as component, added more comments, cleaned unused/deprecated code
