+++ src/components/RestaurantItem.tsx
@@ -2,28 +2,13 @@ import * as React from 'react'
 import { Link } from 'react-router-dom'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { Restaurant } from '../models/models'
+import StarRating from './StarRating'
 
 interface rListItemProps {
   R: Restaurant,
 }
 
 export default function RestaurantItem({ R }: rListItemProps) {
-
-  // Builds array of star icons according to rating
-  const starArr: any = []; // Assigned to type any due to issue with FontAwesomeIcon type
-  let count: number = R.avg_rating; // Temporary store for rating
-  for (let i = 0; i < 5; i++) { // Always gives 5 stars, uses i for unique key
-    if (count >= 1) { // Logic for full stars
-      starArr.push(<FontAwesomeIcon icon={['fas', 'star']} key={i} />)
-    } else if (count > 0 && count < 1) { // Accounts for half stars
-      starArr.push(<FontAwesomeIcon icon="star-half-alt" key={i} />)
-    } else { // Otherwise push empty star
-      starArr.push(<FontAwesomeIcon icon={['far', 'star']} key={i}/>) 
-    }
-    count-- // Decrement rating by 1
-  }
-  
-
   return (
     <li className="card-item">
       <div className="card">
@@ -37,7 +22,7 @@ export default function RestaurantItem({ R }: rListItemProps) {
           <div className="card-title">{ R.title }</div>
 
           <div className="rating">
-            { starArr }
+            <StarRating rating={ R.avg_rating } />
           </div>
 
           <div className="card-text">
