+++ src/components/RestaurantReviewItem.tsx
@@ -3,24 +3,14 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 import { Restaurant, Review } from '../models/models'
 import { Fragment } from 'react';
+import StarRating from './StarRating'
 
 interface reviewsProp {
   review: Review,
 }
 
 export default function RestaurantReviewItem({ review } : reviewsProp) {
-  const starArr: any = []; // Assigned to type any due to issue with FontAwesomeIcon type
-  let count: number = review.rating; // Temporary store for rating
-  for (let i = 0; i < 5; i++) { // Always gives 5 stars, uses i for unique key
-    if (count >= 1) { // Logic for full stars
-      starArr.push(<FontAwesomeIcon icon={['fas', 'star']} key={i} />)
-    } else if (count > 0 && count < 1) { // Accounts for half stars
-      starArr.push(<FontAwesomeIcon icon="star-half-alt" key={i} />)
-    } else { // Otherwise push empty star
-      starArr.push(<FontAwesomeIcon icon={['far', 'star']} key={i} />)
-    }
-    count-- // Decrement rating by 1
-  }
+  
 
   return (
     <div>
@@ -40,8 +30,7 @@ export default function RestaurantReviewItem({ review } : reviewsProp) {
 
       <div>
         <div className="rating">
-          { starArr }
-
+          <StarRating rating={ review.rating }/>
         </div>
 
         <div className="rev-text">
